requirement.txt
python-dotenv==1.1.1
selenium==4.18.1
pyautogui==0.9.54
pyperclip==1.9.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.177.0